清软论坛单元练习/requirements.txt
flask==1.1.2
pyjwt==1.7.1
PyMySQL==1.0.2
scrypt==0.8.18
flask_sqlalchemy==2.5.1
flask_script==2.0.6
flask_migrate==2.5.3
python-dotenv==0.19.0
pyyaml==5.4.1
gunicorn==20.1.0
Jinja2==3.0.3
itsdangerous==1.1.0
werkzeug==0.16.1
selenium==3.141.0
GitPython==3.1.27
kthread==0.2.3
coverage==6.4.4
flasgger==0.9.5